Creating controller of transactions

+++ API.Pagamentos/API.Pagamentos/Repositories/Interfaces/IUserRepository.cs
@@ -5,7 +5,7 @@ namespace API.Pagamentos.Repositories.Interfaces
     public interface IUserRepository
     {
         Task<IEnumerable<User>> GetUsersAsync();
-        Task<User> GetByIdAsync(int? id);
+        Task<User> GetByIdAsync(long? id);
         Task<bool> GetCPF_CNPJUserAsync(string CPF_CNPJ);
         Task<bool> GetEmailUserAsync(string email);
         Task<User> CreateAsync(User user);
