+++ API.Pagamentos/API.Pagamentos/Repositories/UserRepository.cs
@@ -19,7 +19,7 @@ namespace API.Pagamentos.Repositories
             return await _context.Users.ToListAsync();
         }
 
-        public async Task<User> GetByIdAsync(int? id)
+        public async Task<User> GetByIdAsync(long? id)
         {
             return await _context.Users.FindAsync(id) ?? throw new ArgumentNullException();
         }
